Revert using live_server fixture and run test server separately instead
The live_server fixture runs in a separate thread and makes it hard to
load the bakerydemo fixtures. It's also much slower when we start adding
more tests, because the live server needs to be set up for each test
(which includes running migrations etc.).

We don't really need transactional tests anyway. We would like to
emulate a real server as much as possible, which would mean having
complex scenarios that need to be executed in order (e.g. going through
workflows). It would be handy to keep the server's database state
across tests, even though it means that the tests might not be isolated.